Clarify intent of Resolver to make way for new resolver implementation
Since the current `Resolver` is responsible for fetching/storing mappings between hash and string, I've renamed it to `StringResolver`.  This will allow us to implement a `DocumentResolver` that fetches/stores mappings between hash and parsed query documents.

To additionally prep for this change, I've moved the error classes implemented in `Resolver` up a namespace so that they can be shared between `StringResolver` and `DocumentResolver`.